==> notebooks/1_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
import string
import re
import collections
from num2word import word
from wordcloud import WordCloud
from wordcloud import STOPWORDS
from nltk.stem import WordNetLemmatizer
from textblob import TextBlob
import matplotlib.pyplot as plt
import html.parser as parser
from sklearn.feature_extraction.text import TfidfVectorizer
import os

train_file_path = os.path.abspath(os.path.join(os.pardir,'data','train.csv'))
test_file_path = os.path.abspath(os.path.join(os.pardir,'data','test.csv'))

# Import Data Set

Consists of class index 1-4 where 1-World, 2-Sports, 3-Business, 4-Sci/Tech

train = pd.read_csv(train_file_path)
test = pd.read_csv(test_file_path)
data = pd.concat([train,test],axis=0)

#remove html text encoding and #39;
data = pd.concat([train,test],axis=0)
data['Title'] = data['Title'].apply(parser.unescape)
data['Description'] = data['Description'].apply(parser.unescape)
data['Title'] = data['Title'].str.replace("#39;", "\'")
data['Description'] = data['Description'].str.replace("#39;", "\'")

print(train.shape,test.shape,data.shape)

data.head()

# Remove Punctuation and Stopwords 

data['Title'] = data['Title'].apply(lambda s: re.sub(r'[^a-z0-9]',' ',s.lower()))
data['Description'] = data['Description'].apply(lambda s: re.sub(r'[^a-z0-9]',' ',s.lower()))

def convert_num_to_word(words):
    result = []
    for w in words:
        if w.isnumeric():
            result.extend(map(lambda x: x.lower(),word(w).split()))
        else:
            result.append(w)
    return result

data['Title'] = data['Title'].str.split().apply(convert_num_to_word)
data['Description'] = data['Description'].str.split().apply(convert_num_to_word)

def remove_stopword(words):
    result = []
    for word in words:
        if word not in STOPWORDS:
            result.append(word)
    return result

data['Title'] = data['Title'].apply(remove_stopword)
data['Description'] = data['Description'].apply(remove_stopword)

def remove_single_character(words):
    result = []
    for word in words:
        if len(word) > 1:
            result.append(word)
    return result  

data['Title'] = data['Title'].apply(remove_single_character)
data['Description'] = data['Description'].apply(remove_single_character)

# Lemmatization

def lemmatization(words):
    lemmatizer = WordNetLemmatizer()
    result = []
    for word in words:
        result.append(lemmatizer.lemmatize(word))
    return result

data['Title'] = data['Title'].apply(lemmatization)
data['Description'] = data['Description'].apply(lemmatization)

data['Title'] = data['Title'].apply(convert_num_to_word)
data['Description'] = data['Description'].apply(convert_num_to_word)
data['Title'] = data['Title'].apply(remove_stopword)
data['Description'] = data['Description'].apply(remove_stopword)
data['Title'] = data['Title'].apply(remove_single_character)
data['Description'] = data['Description'].apply(remove_single_character)

# Combine Title and Description

data['Documents'] = data['Title'] + data['Description']

data['Word Count'] = data['Documents'].apply(lambda x: len(x))

data.head()

# Convert List to String

data['Documents'] = data['Documents'].apply(lambda x: ' '.join(map(str,x)))
data.drop(['Title','Description'],axis=1,inplace=True)

# Create Features 
### count: noun phrase, noun, verb, adjective

data['Noun Phrases'] = data['Documents'].apply(lambda x: len(TextBlob(x).noun_phrases))

data['Tags'] = data['Documents'].apply(lambda t: collections.Counter(tag for word,tag in TextBlob(t).tags))

# pos tag list https://pythonprogramming.net/part-of-speech-tagging-nltk-tutorial/

data['Noun Count'] = data['Tags'].apply(lambda d: d.get('NN',0)+d.get('NNS',0)+d.get('NNP',0)+d.get('NNPS',0))
data['Adjective Count'] = data['Tags'].apply(lambda d: d.get('JJ',0)+d.get('JJR',0)+d.get('JJS',0))
data['Verb Count'] = data['Tags'].apply(lambda d: d.get('VB',0)+d.get('VBD',0)+d.get('VBG',0)+d.get('VBN',0)+d.get('VBP',0)+d.get('VBZ',0))
data.drop(['Tags'],axis=1,inplace=True)

# Convert Integer Categories to News Categories

#1-World, 2-Sports, 3-Business, 4-Sci/Tech
news_categories = ['','world','sports','business','science_and_technology']

#convert integer categories into corresponding text categories
data['Class Index'] = data['Class Index'].astype(str)
data = data.rename(columns={'Class Index': 'News Category'})

for i in range(len(news_categories)):
    data['News Category'] = data['News Category'].str.replace(str(i), news_categories[i])

# Export Cleaned Data as CSV

data.head()

data_file_path = os.path.abspath(os.path.join(os.pardir,'data','cleaned_AG.csv'))
data_file_path

data.to_csv(data_file_path,index=False)
