==> notebooks/3_tf_idf_features.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
import os

data_file_path = os.path.abspath(os.path.join(os.pardir,'data','cleaned_AG.csv'))
export_file_path = os.path.abspath(os.path.join(os.pardir,'data','tf_idf_training_data.csv'))

# data_file_path = os.path.abspath(os.path.join(os.pardir,'data','cleaned_bing.csv'))
# export_file_path = os.path.abspath(os.path.join(os.pardir,'data','tf_idf_bing_data.csv'))

# Import Datatset
Consists of class index 1-4 where 1-World, 2-Sports, 3-Business, 4-Sci/Tech

data = pd.read_csv(data_file_path)
data.head()

# Feature Extraction: TF-IDF

#top 1000 by term frequency
vectorizer = TfidfVectorizer(max_features=1000)
X = vectorizer.fit_transform(data['Documents'])

denselist = X.todense().tolist()
df = pd.DataFrame(denselist, columns=vectorizer.get_feature_names())

tf_idf = pd.concat([data['News Category'],df],axis=1)

tf_idf.head()

tf_idf.describe()

#export tf idf feature data
tf_idf.to_csv(export_file_path,index=False)
