==> notebooks/4_bag_of_words_features.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
import os

data_file_path = os.path.abspath(os.path.join(os.pardir,'data','cleaned_AG.csv'))
export_file_path = os.path.abspath(os.path.join(os.pardir,'data','bag_of_words_training_data.csv'))

# data_file_path = os.path.abspath(os.path.join(os.pardir,'data','cleaned_bing.csv'))
# export_file_path = os.path.abspath(os.path.join(os.pardir,'data','bag_of_words_bing_data.csv'))

data = pd.read_csv(data_file_path)
data.head()

# Bag of Words

#bag of words
#top 1000 features by term frequency
vectorizer = CountVectorizer(max_features=1000)
X = vectorizer.fit_transform(data['Documents'].values)

#array mapping from feature integer indices to feature name
denselist = X.todense().tolist()
df = pd.DataFrame(denselist, columns=vectorizer.get_feature_names())
bag_of_words = pd.concat([data['News Category'],df],axis=1)
bag_of_words.head()

#export bag of words feature data
bag_of_words.to_csv(export_file_path,index=False)
