==> notebooks/13_word2vec_training_bing.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import matplotlib.cm as cm
import ssl
try:
    _create_unverified_https_context = ssl._create_unverified_context
except AttributeError:
    pass
else:
    ssl._create_default_https_context = _create_unverified_https_context


from gensim.models import Word2Vec
from time import time 
import multiprocessing
import logging  # logger
logging.basicConfig(format="%(levelname)s - %(asctime)s: %(message)s", datefmt= '%H:%M:%S', level=logging.INFO)

from sklearn.manifold import TSNE

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_colwidth', None)

from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.decomposition import PCA

from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, roc_auc_score, roc_curve, auc

data = pd.read_csv('cleaned_AG.csv')
data.head()

# Drop Title & Description

data['Documents'] = data['Documents'].str.split(" ")
data.head()

# Word Embeddings using Word2Vec algorithm

cores = multiprocessing.cpu_count()

# initializing word2vec model
model = Word2Vec(min_count=20,
                     window=2, # window size for context 
                     vector_size=100,  # no of features 
                     sample=6e-5, 
                     alpha=0.03, 
                     min_alpha=0.0007, 
                     negative=20,
                     workers=multiprocessing.cpu_count()-1)

t = time()

# build vocabulary
model.build_vocab(data['Documents'], progress_per=10000)

print('Time to build vocab: {} mins'.format(round((time() - t) / 60, 2)))

t = time()

#train word2vec model 
model.train(data['Documents'], total_examples=model.corpus_count, epochs=30, report_delay=1)

print('Time to train the model: {} mins'.format(round((time() - t) / 60, 2)))

# save model
model.save('model.bin')

# Feature Extraction (word2vec)

Base line number of features = 100

def make_feature_vec(words, model, num_features):
    """
    Average the word vectors for a set of words
    """
    feature_vec = np.zeros((num_features,),dtype="float32")  # pre-initialize (for speed)
    nwords = 0.
    index2word_set = set(model.wv.index_to_key)  # words known to the model

    for word in words:
        if word in index2word_set: 
            nwords = nwords + 1.
            feature_vec = np.add(feature_vec,model.wv[word])
    
    feature_vec = np.divide(feature_vec, nwords)
    return feature_vec


def get_avg_feature_vecs(words, model, num_features):
    """
    Calculate average feature vectors for all headlines 
    """
    counter = 0
    feature_vecs = np.zeros((len(words),num_features), dtype='float32')  # pre-initialize (for speed)
    
    for word in words:
        feature_vecs[counter] = make_feature_vec(word, model, num_features)
        counter = counter + 1
    return feature_vecs

word2vec = get_avg_feature_vecs(data['Documents'], model, 100)

# remove instances in test set that could not be represented as feature vectors
nan_indices = list({x for x,y in np.argwhere(np.isnan(word2vec))})
if len(nan_indices) > 0:
    print('Removing {:d} instances from test set.'.format(len(nan_indices)))
    word2vec = np.delete(word2vec, nan_indices, axis=0)
    word2vec.drop(data.iloc[nan_indices, :].index, axis=0, inplace=True)
    assert word2vec.shape[0] == len(data)

w2v = pd.DataFrame(word2vec)

data.reset_index(drop=True, inplace=True)
w2v.reset_index(drop=True, inplace=True)
#df = pd.concat([df1, df2], axis=1)
w2v = pd.concat([data[['News Category','Word Count','Noun Phrases','Noun Count',
                                         'Adjective Count','Verb Count']],w2v],axis=1)

w2v

# X dataframe 
X = w2v.drop(['News Category'],axis=1) 
# y series
y = w2v['News Category']

X_train,X_test,y_train,y_test = train_test_split(X,y,test_size=0.3)

# Classifications Models

# Random Forest Classifier

from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

from sklearn.ensemble import RandomForestClassifier

# Fit a random forest to the training data, using 100 trees
forest = RandomForestClassifier(n_estimators = 100)

start_time = time()

print("Fitting a random forest to labeled training data...")
#forest = forest.fit(trainDataVecs, train_words['News Category'])
forest = forest.fit(X_train, y_train)

rf_tt = round((time() - start_time) / 60, 2)

print('Time to train Random Forest Model: {} mins'.format(rf_tt))

# start_time = time()

# rf_result = forest.predict(X_test)

# rf_pt = round((time() - start_time) / 60, 2)

# print('Time taken for label prediction using Random Forest model: {} mins'.format(rf_pt))

# rf = round(accuracy_score(y_test, rf_result)*100,3)

# print(classification_report(y_test, rf_result))

# from sklearn.metrics import  confusion_matrix
# import seaborn as sns

# sns.set(font_scale=1.2)
# cof=confusion_matrix(y_test, rf_result)
# cof=pd.DataFrame(cof, index=[i for i in range(1,5)], columns=[i for i in range(1,5)])
# plt.figure(figsize=(7,7))

# sns.heatmap(cof, cmap="BuPu",linewidths=1, annot=True,square=True,cbar=False,fmt='d',
#             xticklabels=['World','Sports','Business','Science'],
#             yticklabels=['World','Sports','Business','Science'])
# plt.xlabel("\nPredicted Class");
# plt.ylabel("Actual Class\n");
# # plt.savefig('W2V Confusion Matrix for News Article Classification using Random Forest.png', bbox_inches='tight')
# plt.title("\nConfusion Matrix for News Article Classification using Random Forest\n");

# Naive Bayes Classsifier

# train a Gaussian Naive Bayes classifier on the training set
from sklearn.naive_bayes import GaussianNB


# instantiate the model
gnb = GaussianNB()

print("Fitting a Naive Bayes to labeled training data...")

start_time = time()

# fit the model
gnb.fit(X_train, y_train)

nb_tt = round((time() - start_time) / 60, 2)

print('Time to train Naive Bayes Model: {} mins'.format(nb_tt))

# start_time = time()

# nb_result = gnb.predict(X_test)

# nb_pt = round((time() - start_time) / 60, 2)

# print('Time taken for label prediction using Naive Bayes model: {} mins'.format(nb_pt))

# nb = round(accuracy_score(y_test, nb_result)*100,3)

# print(classification_report(y_test, nb_result))

# sns.set(font_scale=1.2)
# cof=confusion_matrix(y_test, nb_result)
# cof=pd.DataFrame(cof, index=[i for i in range(1,5)], columns=[i for i in range(1,5)])
# plt.figure(figsize=(7,7))

# sns.heatmap(cof, cmap="BuPu",linewidths=1, annot=True,square=True,cbar=False,fmt='d',
#             xticklabels=['World','Sports','Business','Science'],
#             yticklabels=['World','Sports','Business','Science'])
# plt.xlabel("\nPredicted Class");
# plt.ylabel("Actual Class\n");
# # plt.savefig('W2V Confusion Matrix for News Article Classification using Naive Bayes.png', bbox_inches='tight')
# plt.title("\nConfusion Matrix for News Article Classification using Naive Bayes\n");

# Logistic Regression Classifier

from sklearn.linear_model import LogisticRegression

logit = LogisticRegression()

print("Fitting a Logistic Regression model to labeled training data...")

start_time = time()

logit.fit(X_train,y_train)

lr_tt = round((time() - start_time) / 60, 2)

print('Time to train Logistic Regression Model: {} mins'.format(lr_tt))


# start_time = time()

# lr_result = logit.predict(X_test)

# lr_pt = round((time() - start_time) / 60, 2)

# print('Time taken for label prediction using Logistic Regression model: {} mins'.format(lr_pt))

# lr = round(accuracy_score(y_test, lr_result)*100,3)

# sns.set(font_scale=1.2)
# cof=confusion_matrix(y_test, lr_result)
# cof=pd.DataFrame(cof, index=[i for i in range(1,5)], columns=[i for i in range(1,5)])
# plt.figure(figsize=(7,7))

# sns.heatmap(cof, cmap="BuPu",linewidths=1, annot=True,square=True,cbar=False,fmt='d',
#             xticklabels=['World','Sports','Business','Science'],
#             yticklabels=['World','Sports','Business','Science'])
# plt.xlabel("\nPredicted Class");
# plt.ylabel("Actual Class\n");
# # plt.savefig('W2V Confusion Matrix for News Article Classification using Logistic Regression.png', bbox_inches='tight')
# plt.title("\nConfusion Matrix for News Article Classification using Logistic Regression\n");

# SVM Classifier

from sklearn.svm import SVC

svc = SVC()

print("Fitting a SVM model to labeled training data...")

start_time = time()

svc.fit(X_train,y_train)

svm_tt = round((time() - start_time) / 60, 2)

print('Time to train SVM mode: {} mins'.format(svm_tt))

# start_time = time()

# svm_result = svc.predict(X_test)

# svm_pt = round((time() - start_time) / 60, 2)

# print('Time taken for label prediction using SVM model: {} mins'.format(svm_pt))

# svm = round(accuracy_score(y_test, svm_result)*100,3)

# sns.set(font_scale=1.2)
# cof=confusion_matrix(y_test, svm_result)
# cof=pd.DataFrame(cof, index=[i for i in range(1,5)], columns=[i for i in range(1,5)])
# plt.figure(figsize=(7,7))

# sns.heatmap(cof, cmap="BuPu",linewidths=1, annot=True,square=True,cbar=False,fmt='d',
#             xticklabels=['World','Sports','Business','Science'],
#             yticklabels=['World','Sports','Business','Science'])
# plt.xlabel("\nPredicted Class");
# plt.ylabel("Actual Class\n");
# # plt.savefig('W2V Confusion Matrix for News Article Classification using SVM.png', bbox_inches='tight')
# plt.title("\nConfusion Matrix for News Article Classification using SVM\n");

# Comparison of models (Accuracy)

# sns.set(font_scale=1.2)
# fig = plt.figure()
# ax = fig.add_axes([0,0,1,1])
# Models = ['RandomForest', 'GaussianNB', 'Logistic', 'SVM']
# Accuracy=[rf,nb,lr,svm]
# ax.bar(Models,Accuracy,color=["lightgreen","pink", "skyblue", "lightyellow"]);
# for i in ax.patches:
#     ax.text(i.get_x()+.1, i.get_height()-5.5, str(round(i.get_height(),2))+'%', fontsize=15, color='black')
# plt.title('Comparison of Classification Models trained using word2vec + grammatical + syntax features \n');
# plt.ylabel('Accuracy\n');
# plt.xlabel('\nClassification Models');
# # plt.savefig('W2V Comparison of Classification Models trained using word2vec + grammatical + syntax features.png', bbox_inches='tight')
# plt.show();

# Comparison of models (Training Time)

# sns.set(font_scale=1.2)
# fig = plt.figure()
# ax = fig.add_axes([0,0,1,1])
# Models = ['RandomForest', 'GaussianNB', 'Logistic', 'SVM']
# TrainingTime=[rf_tt,nb_tt,lr_tt,svm_tt]
# ax.bar(Models,TrainingTime,color=["lightyellow","lightgreen", "skyblue", "pink"]);
# for i in ax.patches:
#     ax.text(i.get_x()+.2, i.get_height(), str(round(i.get_height(),2)), fontsize=14, color='black')
# plt.title('Comparison of Different Classification Models Training Time \n');
# plt.ylabel('Training Time (in mins)\n');
# plt.xlabel('\nClassification Models');
# # plt.savefig('W2V Comparison of Different Classification Models Training Time.png', bbox_inches='tight')
# plt.show();

# Comparison of models (Label Classification Time)

# sns.set(font_scale=1.2)
# fig = plt.figure()
# ax = fig.add_axes([0,0,1,1])
# Models = ['RandomForest', 'GaussianNB', 'Logistic', 'SVM']
# PredictTime=[rf_pt,nb_pt,lr_pt,svm_pt]
# ax.bar(Models,PredictTime,color=["lightyellow","lightgreen", "skyblue", "pink"]);
# for i in ax.patches:
#     ax.text(i.get_x()+.2, i.get_height(), str(round(i.get_height(),2)), fontsize=14, color='black')
# plt.title('Comparison of Different Classification Models Label Classification Time \n');
# plt.ylabel('Label Classification Time (in mins)\n');
# plt.xlabel('\nClassification Models');
# # plt.savefig('W2V Comparison of Different Classification Models Label Classification Time.png', bbox_inches='tight')
# plt.show();

# Model Training & Evaluation using Cross-Validation approach

Logistic Regression

# lr = LogisticRegression(solver='newton-cg')

# skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=1)
# lr_acc = []

# for train_index, vaild_index in skf.split(X_train,y_train):
#     x_t, x_v = X_train.iloc[train_index], X_train.iloc[vaild_index]
#     y_t, y_v = y_train.iloc[train_index], y_train.iloc[vaild_index]
#     lr.fit(x_t, y_t)
#     lr_acc.append(lr.score(x_v, y_v))

# print(lr_acc)

# print(np.mean(lr_acc))

Random Forest 

# rf = RandomForestClassifier()

# skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=1)
# rf_acc = []

# for train_index, vaild_index in skf.split(X_train,y_train):
#     x_t, x_v = X_train.iloc[train_index], X_train.iloc[vaild_index]
#     y_t, y_v = y_train.iloc[train_index], y_train.iloc[vaild_index]
#     rf.fit(x_t, y_t)
#     rf_acc.append(rf.score(x_v, y_v))

# print(rf_acc)

# print(np.mean(rf_acc))

Naive Bayes

# gnb = GaussianNB()

# skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=1)
# gnb_acc = []

# for train_index, vaild_index in skf.split(X_train,y_train):
#     x_t, x_v = X_train.iloc[train_index], X_train.iloc[vaild_index]
#     y_t, y_v = y_train.iloc[train_index], y_train.iloc[vaild_index]
#     gnb.fit(x_t, y_t)
#     gnb_acc.append(gnb.score(x_v, y_v))

# print(gnb_acc)

# print(np.mean(gnb_acc))

SVM

# svc = SVC()

# skf = StratifiedKFold(n_splits=2, shuffle=True, random_state=1)
# svc_acc = []

# for train_index, vaild_index in skf.split(X_train,y_train):
#     x_t, x_v = X_train.iloc[train_index], X_train.iloc[vaild_index]
#     y_t, y_v = y_train.iloc[train_index], y_train.iloc[vaild_index]
#     svc.fit(x_t, y_t)
#     svc_acc.append(svc.score(x_v, y_v))

# print(svc_acc)

# print(np.mean(svc_acc))

Model Evaluation

# from sklearn.preprocessing import OneHotEncoder

# y_test = y_test.to_numpy()
# y_pred = lr.predict(X_test)

# lr_acc = accuracy_score(y_test,y_pred)
# lr_recall = recall_score(y_test,y_pred,average='macro')
# lr_precision = precision_score(y_test,y_pred,average='macro')
# lr_f1 = f1_score(y_test,y_pred,average='macro')

# y_pred_roc = OneHotEncoder().fit_transform(y_pred.reshape(-1, 1)).toarray()
# y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
# lr_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

# y_pred = gnb.predict(X_test)

# gnb_acc = accuracy_score(y_test,y_pred)
# gnb_recall = recall_score(y_test_roc,y_pred_roc,average='macro')
# gnb_precision = precision_score(y_test_roc,y_pred_roc,average='macro')
# gnb_f1 = f1_score(y_test_roc,y_pred_roc,average='macro')

# y_pred_roc = OneHotEncoder().fit_transform(y_pred.reshape(-1, 1)).toarray()
# y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
# gnb_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')
# # The multiclass and multilabel cases expect a shape (n_samples, n_classes)

# y_pred = svc.predict(X_test)

# svc_acc = accuracy_score(y_test,y_pred)
# svc_recall = recall_score(y_test,y_pred,average='macro')
# svc_precision = precision_score(y_test,y_pred,average='macro')
# svc_f1 = f1_score(y_test,y_pred,average='macro')

# y_pred_roc = OneHotEncoder().fit_transform(y_pred.reshape(-1, 1)).toarray()
# y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
# svc_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

# y_pred = rf.predict(X_test)

# rf_acc = accuracy_score(y_test,y_pred)
# rf_recall = recall_score(y_test,y_pred,average='macro')
# rf_precision = precision_score(y_test,y_pred,average='macro')
# rf_f1 = f1_score(y_test,y_pred,average='macro')

# y_pred_roc = OneHotEncoder().fit_transform(y_pred.reshape(-1, 1)).toarray()
# y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
# rf_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

# print("Logistic Regression: ",lr_acc,lr_recall,lr_precision,lr_f1,lr_roc)
# print("Naive Bayes: ", gnb_acc,gnb_recall,gnb_precision,gnb_f1,gnb_roc)
# print("Support Vector Machine: ", svc_acc,svc_recall,svc_precision,svc_f1,svc_roc)
# print("Random Forest: ", rf_acc,rf_recall,rf_precision,rf_f1,rf_roc)

# Results

# results = pd.DataFrame([[lr_acc,lr_recall,lr_precision,lr_f1,lr_roc],
#                         [gnb_acc,gnb_recall,gnb_precision,gnb_f1,gnb_roc],
#                         [svc_acc,svc_recall,svc_precision,svc_f1,svc_roc],
#                         [rf_acc,rf_recall,rf_precision,rf_f1,rf_roc]],
#                        columns=['accuracy','recall','precision','fl-score','roc_auc'], 
#                        index=["Logistic Regression","Naive Bayes","Support Vector Machine","Random Forest"])

# results

# results.to_csv('word2vec_results.csv')

# Import Testing Data & W2V Feature Extraction

data = pd.read_csv('cleaned_bing.csv')
data.head()

data['News Category'] = data['News Category'].replace("science","science_and_technology")
data['Documents'] = data['Documents'].str.split(" ")
data.head()

cores = multiprocessing.cpu_count()

# initializing word2vec model
model = Word2Vec(min_count=20,
                     window=2, # window size for context 
                     vector_size=100,  # no of features 
                     sample=6e-5, 
                     alpha=0.03, 
                     min_alpha=0.0007, 
                     negative=20,
                     workers=multiprocessing.cpu_count()-1)

t = time()

# build vocabulary
model.build_vocab(data['Documents'], progress_per=10000)

print('Time to build vocab: {} mins'.format(round((time() - t) / 60, 2)))

t = time()

#train word2vec model 
model.train(data['Documents'], total_examples=model.corpus_count, epochs=30, report_delay=1)

print('Time to train the model: {} mins'.format(round((time() - t) / 60, 2)))

def make_feature_vec(words, model, num_features):
    """
    Average the word vectors for a set of words
    """
    feature_vec = np.zeros((num_features,),dtype="float32")  # pre-initialize (for speed)
    nwords = 0.
    index2word_set = set(model.wv.index_to_key)  # words known to the model

    for word in words:
        if word in index2word_set: 
            nwords = nwords + 1.
            feature_vec = np.add(feature_vec,model.wv[word])
    
    feature_vec = np.divide(feature_vec, nwords)
    return feature_vec


def get_avg_feature_vecs(words, model, num_features):
    """
    Calculate average feature vectors for all headlines 
    """
    counter = 0
    feature_vecs = np.zeros((len(words),num_features), dtype='float32')  # pre-initialize (for speed)
    
    for word in words:
        feature_vecs[counter] = make_feature_vec(word, model, num_features)
        counter = counter + 1
    return feature_vecs

word2vec = get_avg_feature_vecs(data['Documents'], model, 100)

w2v = pd.DataFrame(word2vec)

data.reset_index(drop=True, inplace=True)
w2v.reset_index(drop=True, inplace=True)
w2v = pd.concat([data[['News Category','Word Count','Noun Phrases','Noun Count',
                                         'Adjective Count','Verb Count']],w2v],axis=1)

w2v.dropna(subset=w2v.columns,inplace=True)
w2v.head

# X dataframe 
X_test = w2v.drop(['News Category'],axis=1) 
# y series
y_test = w2v['News Category']

y_test

Model Evaluation

from sklearn.preprocessing import OneHotEncoder

y_test = y_test.to_numpy()
y_pred = logit.predict(X_test)

lr_acc = accuracy_score(y_test,y_pred)
lr_recall = recall_score(y_test,y_pred,average='macro')
lr_precision = precision_score(y_test,y_pred,average='macro')
lr_f1 = f1_score(y_test,y_pred,average='macro')

y_pred_roc = OneHotEncoder().fit(y_test.reshape(-1, 1)).transform(y_pred.reshape(-1,1)).toarray()
y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
lr_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

y_pred = gnb.predict(X_test)

gnb_acc = accuracy_score(y_test,y_pred)
gnb_recall = recall_score(y_test,y_pred,average='macro')
gnb_precision = precision_score(y_test,y_pred,average='macro')
gnb_f1 = f1_score(y_test,y_pred,average='macro')

y_pred_roc = OneHotEncoder().fit(y_test.reshape(-1, 1)).transform(y_pred.reshape(-1,1)).toarray()
y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
gnb_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')
# The multiclass and multilabel cases expect a shape (n_samples, n_classes)

y_pred = svc.predict(X_test)

svc_acc = accuracy_score(y_test,y_pred)
svc_recall = recall_score(y_test,y_pred,average='macro')
svc_precision = precision_score(y_test,y_pred,average='macro')
svc_f1 = f1_score(y_test,y_pred,average='macro')

y_pred_roc = OneHotEncoder().fit(y_test.reshape(-1, 1)).transform(y_pred.reshape(-1,1)).toarray()
y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
svc_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

y_pred = forest.predict(X_test)

rf_acc = accuracy_score(y_test,y_pred)
rf_recall = recall_score(y_test,y_pred,average='macro')
rf_precision = precision_score(y_test,y_pred,average='macro')
rf_f1 = f1_score(y_test,y_pred,average='macro')

y_pred_roc = OneHotEncoder().fit(y_test.reshape(-1, 1)).transform(y_pred.reshape(-1,1)).toarray()
y_test_roc = OneHotEncoder().fit_transform(y_test.reshape(-1, 1)).toarray()
rf_roc = roc_auc_score(y_test_roc,y_pred_roc,multi_class='ovo')

print("Logistic Regression: ",lr_acc,lr_recall,lr_precision,lr_f1,lr_roc)
print("Naive Bayes: ", gnb_acc,gnb_recall,gnb_precision,gnb_f1,gnb_roc)
print("Support Vector Machine: ", svc_acc,svc_recall,svc_precision,svc_f1,svc_roc)
print("Random Forest: ", rf_acc,rf_recall,rf_precision,rf_f1,rf_roc)

# Results

results = pd.DataFrame([[lr_acc,lr_recall,lr_precision,lr_f1,lr_roc],
                        [gnb_acc,gnb_recall,gnb_precision,gnb_f1,gnb_roc],
                        [svc_acc,svc_recall,svc_precision,svc_f1,svc_roc],
                        [rf_acc,rf_recall,rf_precision,rf_f1,rf_roc]],
                       columns=['accuracy','recall','precision','fl-score','roc_auc'], 
                       index=["Logistic Regression","Naive Bayes","Support Vector Machine","Random Forest"])

results

results.to_csv('word2vec_bing_results.csv')
