==> notebooks/14_Result_Comparison.ipynb <==
import pandas as pd
pd.set_option('precision',2)

bow = pd.read_csv('bag_of_words_results.csv',index_col=0)
tf_idf = pd.read_csv('tf-idf_results.csv',index_col=0)
word2vec = pd.read_csv('word2vec_results.csv',index_col=0)

bow

tf_idf

word2vec

pd.concat([bow['accuracy'].rename('bow_accuracy'),
           tf_idf['accuracy'].rename('tfidf_accuracy'),
           word2vec['accuracy'].rename('word2vec_accuracy')], axis=1)

pd.concat([bow['recall'].rename('bow_recall'),
           tf_idf['recall'].rename('tfidf_recall'),
           word2vec['recall'].rename('word2vec_recall')], axis=1)

pd.concat([bow['precision'].rename('bow_precision'),
           tf_idf['precision'].rename('tfidf_precision'),
           word2vec['precision'].rename('word2vec_precision')], axis=1)

pd.concat([bow['fl-score'].rename('bow_f1-score'),
           tf_idf['fl-score'].rename('tfidf_f1-score'),
           word2vec['fl-score'].rename('word2vec_f1-score')], axis=1)

pd.concat([bow['roc_auc'].rename('bow_roc_auc'),
           tf_idf['roc_auc'].rename('tfidf_roc_auc'),
           word2vec['roc_auc'].rename('word2vec_roc_auc')], axis=1)
